feat: get cookie from BFF
