+++ package.json
@@ -58,6 +58,7 @@
     "axios": "^1.4.0",
     "chalk": "^4.1.2",
     "conf": "^10.2.0",
+    "cookie-parser": "^1.4.6",
     "degit": "^2.8.4",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
@@ -75,6 +76,7 @@
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/npm": "^11.0.0",
     "@types/chai": "4.3.5",
+    "@types/cookie-parser": "^1.4.5",
     "@types/degit": "^2.8.4",
     "@types/express": "^4.17.17",
     "@types/fs-extra": "^11.0.2",
