--- src/services/affinidi/auth/providers/index.ts
@@ -1,4 +1,0 @@
-export { OryAuthenticatorPKCE } from './ory-pkce'
-export { OryAuthenticator } from './ory'
-
-export { AuthProvider, AuthProviderConfig } from './types'
