+++ src/services/affinidi/auth/providers/types.ts
@@ -6,7 +6,6 @@ export interface AuthProvider {
 }
 
 export type AuthProviderConfig = {
-  clientId: string
   host: string
   logger: LoggerAdapter
 }
