+++ src/services/affinidi/iam/index.ts
@@ -69,13 +69,6 @@ class IAMService {
         },
       })
 
-      //TODO Remove this backwards compatibility
-      if (!data.principalType) {
-        const principalComponents = data.principalId.split('/')
-        data.principalId = principalComponents[1]
-        data.principalType = principalComponents[0]
-      }
-
       return data
     } catch (error) {
       handleServiceError(error)
